define single entry-point for component loading
